moving FileBPlusTree and tests containing it to jvm dir
